Separate path folding from tail folding logic
